move IncreaseLockedMemoryQuota to bpfutils
